Fix: 웹소켓이 모든 페이지에 적용되지 않는 부분 수정
